Make Blink more standard conformant

+++ blink/close.c
@@ -96,7 +96,7 @@ static int SysCloseRangeCloexec(struct Machine *m, u32 first, u32 last) {
   for (e = dll_first(m->system->fds.list); e;
        e = dll_next(m->system->fds.list, e)) {
     fd = FD_CONTAINER(e);
-    if (first <= fd->fildes && fd->fildes <= last) {
+    if (first <= (u32)fd->fildes && (u32)fd->fildes <= last) {
       if (~fd->oflags & O_CLOEXEC) {
         fd->oflags |= O_CLOEXEC;
         fcntl(fd->fildes, F_SETFD, FD_CLOEXEC);
@@ -122,7 +122,7 @@ int SysCloseRange(struct Machine *m, u32 first, u32 last, u32 flags) {
   for (fds = 0, e = dll_first(m->system->fds.list); e; e = e2) {
     fd = FD_CONTAINER(e);
     e2 = dll_next(m->system->fds.list, e);
-    if (first <= fd->fildes && fd->fildes <= last) {
+    if (first <= (u32)fd->fildes && (u32)fd->fildes <= last) {
       dll_remove(&m->system->fds.list, e);
       dll_make_last(&fds, e);
     }
